+++ src/components/results/ResultModal.js
